requirements.txt
PyQt6
openai
pyaudio
python-dotenv
pyperclip
pynput
faster-whisper
soundcard
numpy
soundfile
